[UPD] 更新模版为hap4.0版本
