+++ README.md
@@ -1,9 +1,7 @@
-# To customize a template
-you need to push the template code to this git repository.
+项目技术人员应此处描述一些项目基本信息，及注意事项：
 
-Please make sure the following file exists.
-+ **gitlab-ci.yml**. (Refer to [GitLab Documentation](https://docs.gitlab.com/ee/ci/yaml/))
-+ **Dockerfile**. (Refer to [Dockerfile reference](https://docs.docker.com/engine/reference/builder/))
-+ **Chart** setting directory. (Refer to [helm](https://github.com/kubernetes/helm))
+* 1，使用的技术介绍
 
-Finally, removing or re-editing this **README.md** file to make it useful.
+* 2，主要功能
+
+* 3，配置信息
