+++ charts/model-service/templates/ingress.yaml
@@ -1,4 +1,4 @@
-{{- if .Values.ingress.enable }}
+{{- if .Values.ingress.enabled }}
 apiVersion: extensions/v1beta1
 kind: Ingress
 metadata:
