+++ charts/model-service/templates/service.yaml
@@ -1,4 +1,4 @@
-{{- if .Values.service.enable }}
+{{- if .Values.service.enabled }}
 apiVersion: v1
 kind: Service
 metadata:
