+++ charts/model-service/values.yaml
@@ -1,122 +1,47 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+replicaCount: 1
 
-#
-# List of comma-separated packages that start with or equal this string
-# will cause a security exception to be thrown when
-# passed to checkPackageAccess unless the
-# corresponding RuntimePermission ("accessClassInPackage."+package) has
-# been granted.
-package.access=sun.,org.apache.catalina.,org.apache.coyote.,org.apache.jasper.,org.apache.tomcat.
-  #
-  # List of comma-separated packages that start with or equal this string
-  # will cause a security exception to be thrown when
-  # passed to checkPackageDefinition unless the
-  # corresponding RuntimePermission ("defineClassInPackage."+package) has
-  # been granted.
-  #
-  # by default, no packages are restricted for definition, and none of
-  # the class loaders supplied with the JDK call checkPackageDefinition.
-  #
-package.definition=sun.,java.,org.apache.catalina.,org.apache.coyote.,\
-org.apache.jasper.,org.apache.naming.,org.apache.tomcat.
+image:
+  repository: registry.cn-hangzhou.aliyuncs.com/{{group.name}}/{{service.code}}
+  pullPolicy: IfNotPresent
 
-  #
-  #
-  # List of comma-separated paths defining the contents of the "common"
-  # classloader. Prefixes should be used to define what is the repository type.
-  # Path may be relative to the CATALINA_HOME or CATALINA_BASE path or absolute.
-  # If left as blank,the JVM system loader will be used as Catalina's "common"
-  # loader.
-  # Examples:
-  #     "foo": Add this folder as a class repository
-  #     "foo/*.jar": Add all the JARs of the specified folder as class
-  #                  repositories
-  #     "foo/bar.jar": Add bar.jar as a class repository
-  #
-  # Note: Values are enclosed in double quotes ("...") in case either the
-  #       ${catalina.base} path or the ${catalina.home} path contains a comma.
-  #       Because double quotes are used for quoting, the double quote character
-  #       may not appear in a path.
-common.loader="${catalina.base}/lib","${catalina.base}/lib/*.jar","${catalina.home}/lib","${catalina.home}/lib/*.jar"
+preJob:
+  timeout: 300
+  image: registry.cn-hangzhou.aliyuncs.com/choerodon-tools/dbtool:0.6.5
+  preInitDB:
+    enabled: true
+    datasource:
+      url: jdbc:mysql://localhost:3306/hap_dev?useUnicode=true&characterEncoding=utf-8&useSSL=false
+      username: hap_dev
+      password: hap_dev
 
-  #
-  # List of comma-separated paths defining the contents of the "server"
-  # classloader. Prefixes should be used to define what is the repository type.
-  # Path may be relative to the CATALINA_HOME or CATALINA_BASE path or absolute.
-  # If left as blank, the "common" loader will be used as Catalina's "server"
-  # loader.
-  # Examples:
-  #     "foo": Add this folder as a class repository
-  #     "foo/*.jar": Add all the JARs of the specified folder as class
-  #                  repositories
-  #     "foo/bar.jar": Add bar.jar as a class repository
-  #
-  # Note: Values may be enclosed in double quotes ("...") in case either the
-  #       ${catalina.base} path or the ${catalina.home} path contains a comma.
-  #       Because double quotes are used for quoting, the double quote character
-  #       may not appear in a path.
-server.loader=
+service:
+  enable: false
+  type: ClusterIP
+  port: 8080
+  name: {{service.code}}
 
-  #
-  # List of comma-separated paths defining the contents of the "shared"
-  # classloader. Prefixes should be used to define what is the repository type.
-  # Path may be relative to the CATALINA_BASE path or absolute. If left as blank,
-  # the "common" loader will be used as Catalina's "shared" loader.
-  # Examples:
-  #     "foo": Add this folder as a class repository
-  #     "foo/*.jar": Add all the JARs of the specified folder as class
-  #                  repositories
-  #     "foo/bar.jar": Add bar.jar as a class repository
-  # Please note that for single jars, e.g. bar.jar, you need the URL form
-  # starting with file:.
-  #
-  # Note: Values may be enclosed in double quotes ("...") in case either the
-  #       ${catalina.base} path or the ${catalina.home} path contains a comma.
-  #       Because double quotes are used for quoting, the double quote character
-  #       may not appear in a path.
-shared.loader=
+ingress:
+  enable: false
+  host: choerodon.com.cn
 
-  # Default list of JAR files that should not be scanned using the JarScanner
-  # functionality. This is typically used to scan JARs for configuration
-  # information. JARs that do not contain such information may be excluded from
-  # the scan to speed up the scanning process. This is the default list. JARs on
-  # this list are excluded from all scans. The list must be a comma separated list
-  # of JAR file names.
-  # The list of JARs to skip may be over-ridden at a Context level for individual
-  # scan types by configuring a JarScanner with a nested JarScanFilter.
-  # The JARs listed below include:
-  # - Tomcat Bootstrap JARs
-  # - Tomcat API JARs
-  # - Catalina JARs
-  # - Jasper JARs
-  # - Tomcat JARs
-  # - Common non-Tomcat JARs
-  # - Test JARs (JUnit, Cobertura and dependencies)
-tomcat.util.scan.StandardJarScanFilter.jarsToSkip=*.jar
+metrics:
+  path: /actuator/prometheus
+  group: spring-boot
 
-  # Default list of JAR files that should be scanned that overrides the default
-  # jarsToSkip list above. This is typically used to include a specific JAR that
-  # has been excluded by a broad file name pattern in the jarsToSkip list.
-  # The list of JARs to scan may be over-ridden at a Context level for individual
-  # scan types by configuring a JarScanner with a nested JarScanFilter.
-tomcat.util.scan.StandardJarScanFilter.jarsToScan=\
-log4j-core*.jar,log4j-taglib*.jar,log4javascript*.jar,slf4j-taglib*.jar
+env:
+  open:
+    SPRING_DATASOURCE_URL: jdbc:mysql://localhost:3306/hap_dev?useUnicode=true&characterEncoding=utf-8&useSSL=false
+    SPRING_DATASOURCE_USERNAME: hap_dev
+    SPRING_DATASOURCE_PASSWORD: hap_dev
+    SPRING_REDIS_HOST: localhost
+    REDIS_DB: 6379
+    SPRING_REDIS_DATABASE: 10
 
-  # String cache configuration.
-tomcat.util.buf.StringCache.byte.enabled=true
-#tomcat.util.buf.StringCache.char.enabled=true
-#tomcat.util.buf.StringCache.trainThreshold=500000
-#tomcat.util.buf.StringCache.cacheSize=5000
+logs:
+  parser: spring-boot
+
+resources:
+  limits:
+    memory: 2Gi
+  requests:
+    memory: 1Gi
