--- src/main/docker/Dockerfile
@@ -1,3 +1,0 @@
-FROM registry.cn-hangzhou.aliyuncs.com/choerodon-tools/javabase:0.6.0
-COPY hap-demo.jar /{{service.code}}.jar
-ENTRYPOINT exec java $JAVA_OPTS  -jar /{{service.code}}.jar
