--- src/main/docker/entrypoint.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-java -Xms512M -Xmx1024M  -jar hap-demo.jar
