--- src/main/java/hbi/core/grid/mapper/GridDemoMapper.java
@@ -1,8 +1,0 @@
-package hbi.core.grid.mapper;
-
-import hbi.core.grid.dto.GridDemo;
-import com.hand.hap.mybatis.common.Mapper;
-
-public interface GridDemoMapper extends Mapper<GridDemo> {
-
-}
