--- src/main/java/hbi/core/grid/service/IGridDemoService.java
@@ -1,9 +1,0 @@
-package hbi.core.grid.service;
-
-import com.hand.hap.core.ProxySelf;
-import hbi.core.grid.dto.GridDemo;
-import com.hand.hap.system.service.IBaseService;
-
-public interface IGridDemoService extends IBaseService<GridDemo>, ProxySelf<IGridDemoService> {
-
-}
