+++ src/main/java/io/choerodon/grid/dto/GridDemo.java
@@ -1,0 +1,128 @@
+package io.choerodon.grid.dto;
+
+import io.choerodon.mybatis.annotation.EnableExtensionAttribute;
+import io.choerodon.mybatis.annotation.MultiLanguage;
+import io.choerodon.mybatis.annotation.MultiLanguageField;
+import io.choerodon.mybatis.common.query.Where;
+import io.choerodon.mybatis.entity.BaseDTO;
+import org.hibernate.validator.constraints.Length;
+
+import javax.validation.constraints.NotEmpty;
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OrderBy;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+import javax.validation.constraints.NotNull;
+import java.util.Date;
+
+@MultiLanguage
+@EnableExtensionAttribute
+@Table(name = "hap_grid_demo_b")
+public class GridDemo extends BaseDTO {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Where
+    @Column
+    @OrderBy("DESC")
+    private Long id;
+
+    @NotEmpty
+    @Where
+    @Length(max = 50)
+    @Column
+    @MultiLanguageField
+    @OrderBy
+    private String name;
+
+    @NotNull
+    @Column
+    private Long age;
+
+    @Column
+    @NotEmpty
+    @Length(max = 1)
+    private String sex;
+
+    @Column
+    @Where
+    @MultiLanguageField
+    @Length(max = 255)
+    private String description;
+
+    @Column
+    private String positionId;
+
+    @Transient
+    private String positionName;
+
+    @Column
+    private Date birth;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Long getAge() {
+        return age;
+    }
+
+    public void setAge(Long age) {
+        this.age = age;
+    }
+
+    public String getSex() {
+        return sex;
+    }
+
+    public void setSex(String sex) {
+        this.sex = sex;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getPositionId() {
+        return positionId;
+    }
+
+    public void setPositionId(String positionId) {
+        this.positionId = positionId;
+    }
+
+    public String getPositionName() {
+        return positionName;
+    }
+
+    public void setPositionName(String positionName) {
+        this.positionName = positionName;
+    }
+
+    public Date getBirth() {
+        return birth;
+    }
+
+    public void setBirth(Date birth) {
+        this.birth = birth;
+    }
+}
