+++ src/main/java/io/choerodon/grid/service/IGridDemoService.java
@@ -1,0 +1,9 @@
+package io.choerodon.grid.service;
+
+import io.choerodon.hap.core.ProxySelf;
+import io.choerodon.mybatis.service.IBaseService;
+import io.choerodon.grid.dto.GridDemo;
+
+public interface IGridDemoService extends IBaseService<GridDemo>, ProxySelf<IGridDemoService> {
+
+}
