+++ src/main/resources/io/choerodon/grid/mapper/GridDemoMapper.xml
@@ -1,0 +1,91 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
+<mapper namespace="io.choerodon.grid.mapper.GridDemoMapper">
+    <resultMap id="BaseResultMap" type="io.choerodon.grid.dto.GridDemo" extends="io.choerodon.mybatis.mapper.StdMapper.STD">
+        <result column="ID" property="id" jdbcType="DECIMAL"/>
+        <result column="NAME" property="name" jdbcType="VARCHAR"/>
+        <result column="AGE" property="age" jdbcType="DECIMAL"/>
+        <result column="SEX" property="sex" jdbcType="VARCHAR"/>
+        <result column="DESCRIPTION" property="description" jdbcType="VARCHAR"/>
+        <result column="POSITION_ID" property="positionId" jdbcType="DECIMAL"/>
+        <result column="POSITION_NAME" property="positionName" jdbcType="VARCHAR"/>
+        <result column="BIRTH" property="birth" javaType="DATE"/>
+    </resultMap>
+
+    <select id="selectGridXml" resultMap="BaseResultMap" parameterType="java.util.Map">
+        SELECT
+        HGDB.ID,
+        HGDT.NAME,
+        HGDB.AGE,
+        HGDB.SEX,
+        HGDT.DESCRIPTION,
+        HGDB.POSITION_ID,
+        HOPT.NAME AS POSITION_NAME,
+        HGDB.BIRTH,
+        HGDB.OBJECT_VERSION_NUMBER
+        FROM  HAP_GRID_DEMO_B HGDB
+        LEFT OUTER JOIN HAP_GRID_DEMO_TL HGDT ON (
+            HGDB.ID = HGDT.ID
+            AND HGDT.LANG = #{request.locale,jdbcType=VARCHAR,javaType=java.lang.String}
+        )
+        LEFT JOIN HR_ORG_POSITION_TL HOPT ON (
+            HOPT.POSITION_ID = HGDB.POSITION_ID
+            AND HOPT.LANG = #{request.locale,jdbcType=VARCHAR,javaType=java.lang.String}
+        )
+        <where>
+            <if test="id !=null ">
+                AND HGDB.ID =  #id
+            </if>
+            <if test="name !=null ">
+                AND HGDT.NAME LIKE CONCAT('%', CONCAT(#{name,jdbcType=VARCHAR}, '%'))
+            </if>
+            <if test="description!=null">
+                AND HGDT.DESCRIPTION LIKE CONCAT('%', CONCAT(#{description,jdbcType=VARCHAR}, '%'))
+            </if>
+        </where>
+        <if test="__sort_column">
+            ORDER BY ${__sort_column} ${__order_flag}
+        </if>
+    </select>
+
+    <select id="selectGrid" resultMap="BaseResultMap" parameterType="io.choerodon.grid.dto.GridDemo">
+        SELECT
+        HGDB.ID,
+        HGDT.NAME,
+        HGDB.AGE,
+        HGDB.SEX,
+        HGDT.DESCRIPTION,
+        HGDB.POSITION_ID,
+        HOPT.NAME AS POSITION_NAME,
+        HGDB.BIRTH,
+        HGDB.OBJECT_VERSION_NUMBER
+        FROM  HAP_GRID_DEMO_B HGDB
+        LEFT OUTER JOIN HAP_GRID_DEMO_TL HGDT ON (
+        HGDB.ID = HGDT.ID
+        AND HGDT.LANG = #{request.locale,jdbcType=VARCHAR,javaType=java.lang.String}
+        )
+        LEFT JOIN HR_ORG_POSITION_TL HOPT ON (
+        HOPT.POSITION_ID = HGDB.POSITION_ID
+        AND HOPT.LANG = #{request.locale,jdbcType=VARCHAR,javaType=java.lang.String}
+        )
+        <where>
+            <if test="id !=null ">
+                AND HGDB.ID =  #id
+            </if>
+            <if test="name !=null ">
+                AND HGDT.NAME LIKE CONCAT('%', CONCAT(#{name,jdbcType=VARCHAR}, '%'))
+            </if>
+            <if test="description!=null">
+                AND HGDT.DESCRIPTION LIKE CONCAT('%', CONCAT(#{description,jdbcType=VARCHAR}, '%'))
+            </if>
+        </where>
+        <if test="sortname !=null">
+            <bind name="_colName" value="@io.choerodon.mybatis.util.OGNL@unCamel(sortname)"/>
+            ORDER BY HGDB.${_colName} ${sortorder}
+        </if>
+        <if test="sortname ==null">
+            ORDER BY HGDB.ID DESC
+        </if>
+    </select>
+</mapper>
